rename: my-page -> user 변경
