+++ app/components/user/CategoryChart.tsx
